+++ apps/www/src/modules/about/page/index.tsx
@@ -2,7 +2,7 @@ import { Separator } from "@components/separator";
 import AboutUsHeading from "../widgets/heading";
 import Content from "../widgets/content";
 import Team from "../widgets/team";
-import ServiceSection from "../widgets/services-section";
+// import ServiceSection from "../widgets/services-section";
 import FooterSetup from "@utilities/footer-setup";
 
 const AboutPage = () => {
@@ -15,7 +15,7 @@ const AboutPage = () => {
       <Team container />
       <Separator className="mt-28 mb-8" container />
       {/* Here should be DeftDodo Banner [DDT-16] */}
-      <ServiceSection container />
+      {/* <ServiceSection container /> */}
       <FooterSetup variant="complex" />
     </>
   );