+++ apps/www/src/modules/common/types/server.d.ts
@@ -1,0 +1,16 @@
+export interface ServerResponse<TServerAttributes> {
+  data: ServerData<TServerAttributes>;
+  meta: object;
+}
+
+export interface ServerData<TServerAttributes> {
+  id: number;
+  attributes: TServerAttributes;
+}
+
+export interface BaseServerAttributes {
+  createAt: string;
+  updatedAt: string;
+  publishedAt: string;
+  locale: string;
+}