+++ apps/www/src/modules/home/components/hero-video.tsx
@@ -1,13 +1,25 @@
 "use client";
 
 import Image from "next/image";
-import videoOverlay from "../assets/video-thubnail.webp";
 import { useState } from "react";
 import { Dialog, DialogContent } from "@components/dialog";
 import { Play } from "lucide-react";
-const HeroVideo = () => {
+import { MediaResponse } from "@/modules/common/types/common";
+import { AppConfig } from "@/config/app.config";
+
+type HeroVideoProps = {
+  youtube_video_hash?: string;
+  overlay?: MediaResponse;
+};
+
+const HeroVideo: React.FC<HeroVideoProps> = ({
+  youtube_video_hash,
+  overlay,
+}) => {
   const [playing, setPlaying] = useState(false);
 
+  if (!overlay) return null;
+  const imageSize = overlay?.data.attributes.formats.large;
   return (
     <>
       <div
@@ -16,23 +28,26 @@ const HeroVideo = () => {
       >
         <Play className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 size-16 text-white-16 group-hover:text-white transition-all duration-300 ease-in-out group-hover:size-20 fill-white-16 group-hover:fill-white-56" />
         <Image
-          src={videoOverlay}
-          alt="Hero Carousel"
-          width={1400}
-          height={650}
+          src={`${AppConfig.strapi.url}${imageSize.url}`}
+          alt={overlay?.data.attributes.alternativeText || "Hero Carousel"}
+          width={imageSize.width}
+          height={imageSize.height}
           className="w-full h-auto rounded-xl"
           priority
         />
       </div>
+
       {playing && (
         <Dialog open={playing} onOpenChange={setPlaying}>
           <DialogContent className="p-1 bg-primary w-[85vw] max-w-[1200px]">
-            <div className="embed-container">
-              <iframe
-                src="https://www.youtube.com/embed/SVU02zJ_R4o?autoplay=1&loop=1"
-                allow="autoplay;"
-              ></iframe>
-            </div>
+            {!!youtube_video_hash && (
+              <div className="embed-container">
+                <iframe
+                  src={`https://www.youtube.com/embed/${youtube_video_hash}?autoplay=1&loop=1`}
+                  allow="autoplay;"
+                ></iframe>
+              </div>
+            )}
           </DialogContent>
         </Dialog>
       )}