+++ apps/www/src/modules/home/page/index.tsx
@@ -1,52 +1,38 @@
-import React from "react";
 import HeroCarousel from "../components/hero-carousel";
 import HeroSection from "../widgets/hero";
-import Services from "../widgets/services";
 import { Separator } from "@components/separator";
-import Heading from "../widgets/heading";
 import Clients from "../widgets/clients";
 import AboutUs from "../widgets/about-us";
 import { Projects } from "../widgets/projects";
 import BannerMarquee from "@/modules/common/components/banner-marquee";
 import Experience from "@/modules/common/components/experience";
 import FooterSetup from "@/modules/common/utilities/footer-setup";
+import { getHomePageData } from "../actions/home-page";
 import KeyFactsSection from "../widgets/key-facts";
 import FeaturedServices from "../widgets/featured-services";
-import { type ServiceCardProps } from "@/modules/services/components/service-card";
 import Testimonials from "../widgets/testimonials";
 import { type TestimonialCardProps } from "../components/testimonial-card";
+import OurProcess from "../widgets/our-process";
+import { ProcessItem } from "../components/our-process-animated";
 
-// this is a reference array of featured services
-
-const services: ServiceCardProps[] = [
+const process: ProcessItem[] = [
   {
-    icon: "BarChart2",
-    title: "CRO Audit",
+    number: "01",
+    title: "Research & Plan",
     description:
-      "A comprehensive assessment of your Shopify e-commerce store's performance with the goal of identifying opportunities to improve its conversion rates. Conversion rate optimization focuses on increasing the percentage of website.",
-    slug: "cro-audit",
-    variant: "primary",
+      "Design process is critical for gathering information, requirements, and other data you need in order to make informed decisions later.",
   },
   {
-    icon: "Layout",
-    title: "Theme Customization",
+    number: "02",
+    title: "Mock Up",
     description:
-      "Refers to the process of modifying and tailoring the design, layout, features, and functionality of a pre-existing Shopify theme to better align with a store owner's brand, vision, and specific business needs.",
-    slug: "theme-customization",
+      "Creating a mock up is the step of the design process most people recognize — it’s the most visual part of the process In-depth sketching and brainstorming.",
   },
   {
-    icon: "TrendingUp",
-    title: "SEO & Digital Marketing Integration",
+    number: "03",
+    title: "Build",
     description:
-      "Involve strategies and techniques to enhance the visibility, reach, and effectiveness of your Shopify e-commerce store in search engines and various online marketing channels.",
-    slug: "seo-digital-marketing-integration",
-  },
-  {
-    icon: "Zap",
-    title: "Performance Optimization",
-    description:
-      "Improving the speed, responsiveness, and overall efficiency of your Shopify e-commerce store. The goal of performance optimization is to create a seamless and enjoyable user experience.",
-    slug: "performance-optimization",
+      "Implement the solution (often with the help of other professionals like programmers, printers, or manufacturers). Revise the solution as technical issues.",
   },
 ];
 
@@ -69,52 +55,67 @@ const testimonials: TestimonialCardProps[] = [
   }
 ];
 
-const HomePage = () => {
+const HomePage = async () => {
+  const data = await getHomePageData();
+  const pageContent = data.data.attributes.pageContent;
+
   return (
     <>
-      <HeroSection
-        container
-        headingBold="We are creative"
-        headingThin="Agency"
-        description="Deft Dodo is a design & develop agency of the future with 120+ successful projects in their portfolio."
-        cta="Book a demo"
-        href="/contact"
-      />
-      <Heading container />
-      <Separator className="my-12" container />
-      <HeroCarousel container />
-      <Clients container className="mt-28" />
-      <Separator className="my-28" />
-      <AboutUs
+      {pageContent.map((content, index) => {
+        switch (content.__component) {
+          case "home.heading":
+            return <HeroSection key={index} {...content} />;
+          case "home.heading-carousel":
+            return <HeroCarousel key={index} {...content} />;
+          case "home.clients":
+            return <Clients key={index} {...content} />;
+          case "home.about-us":
+            return <AboutUs key={index} {...content} />;
+          case "home.projects":
+            return <Projects key={index} {...content} />;
+          case "home.services":
+            return <FeaturedServices key={index} {...content} />;
+          case "home.key-facts":
+            return <KeyFactsSection key={index} {...content} />;
+          case "common.separator":
+            return (
+              <Separator
+                key={index}
+                container={content.with_container}
+                style={{
+                  marginBottom: `${content.margin_button}px`,
+                  marginTop: `${content.margin_top}px`,
+                }}
+              />
+            );
+          case "common.spacer":
+            return (
+              <div
+                style={{
+                  height: `${content.height}px`,
+                }}
+              />
+            );
+          default:
+            return null;
+        }
+      })}
+      <OurProcess
         id={1}
-        __component="home.about-us"
+        __component="home.our-process"
         with_container={true}
         heading={{
           id: 1,
-          title: "About Us",
-          sub_title: "We are a creative agency",
+          title: "Our Process",
         }}
-        content="Deft Dodo is a design & develop agency of the future with 120+ successful projects in their portfolio."
+        items={process}
       />
       <Separator className="my-28" />
-      <KeyFactsSection container />
-      <Separator className="mt-28 mb-10" />
-      <Projects container />
-      <Separator className="mt-28 mb-2" container />
-      <FeaturedServices
-        container
-        heading="Know our services"
-        subHeading="Featured Services"
-        description="We offer a range of services to help you grow your business."
-        services={services}
-        viewAllLink="/services"
-      />
       <Testimonials
         container
         heading="Testimonials"
         testimonials={testimonials}
       />
-      <Services container />
       <Separator className="mt-24 mb-36" container />
       <BannerMarquee />
       <Experience />