+++ apps/www/src/modules/home/widgets/hero.tsx
@@ -2,23 +2,29 @@ import { type HeroBannerProps } from "../components/hero-banner";
 import HeroBanner from "../components/hero-banner";
 import ParticlesBackground from "../components/particles-background";
 
-const HeroSection: React.FC<HeroBannerProps> = ({ container, headingBold, headingThin, description, cta, href }: HeroBannerProps) => {
-
-    return (
-        <section className="relative">
-            <HeroBanner
-                container={container}
-                headingBold={headingBold}
-                headingThin={headingThin}
-                description={description}
-                cta={cta}
-                href={href}
-            />
-            <div className="absolute inset-0 h-full pointer-events-none">
-                <ParticlesBackground />
-            </div>
-        </section>
-    )
+const HeroSection: React.FC<HeroBannerProps> = ({
+  with_container: container,
+  heading_bold: headingBold,
+  heading_thing: headingThin,
+  description,
+  cta,
+  href,
+}: HeroBannerProps) => {
+  return (
+    <section className="relative">
+      <HeroBanner
+        with_container={container}
+        heading_bold={headingBold}
+        heading_thing={headingThin}
+        description={description}
+        cta={cta}
+        href={href}
+      />
+      <div className="absolute inset-0 h-full pointer-events-none">
+        <ParticlesBackground />
+      </div>
+    </section>
+  );
 };
 
 export default HeroSection;