+++ apps/www/src/modules/services/components/service-card.tsx
@@ -1,24 +1,19 @@
 "use client";
 
-import { Icon, IconName } from "@/modules/common/components/icon";
+import { Icon } from "@/modules/common/components/icon";
+import { Services } from "@/modules/common/types/common";
+import { iconify } from "@/modules/common/utilities/icon.utility";
 import { Button } from "@components/button";
 import useAppData from "@hooks/useAppData";
 
-export type ServiceCardProps = {
-  title: string;
-  description: string;
-  slug: string;
-  icon: IconName;
+export type ServiceCardProps = Services & {
   variant?: "default" | "primary";
 };
 
 type ButtonVariant = "default" | "outline";
 
 const ServiceCard: React.FC<ServiceCardProps> = ({
-  title,
-  description,
-  icon,
-  slug,
+  attributes: { title, cta, icon, description, slug },
   variant = "default",
 }) => {
   const { locale } = useAppData();
@@ -33,25 +28,42 @@ const ServiceCard: React.FC<ServiceCardProps> = ({
       text: "text-white",
       buttonVariant: "default" as ButtonVariant,
     },
-  }
+  };
 
   return (
-    <section className={settings[variant].background + " p-8 rounded-lg space-y-8 h-full"}>
-      <Icon name={icon} className={settings[variant].text + " size-16"} />
-      <div className="space-y-3">
-        <h3 className="font-bold text-xl xl:text-2xl text-white sm:min-h-14 lg:min-h-16">
-          {title}
-        </h3>
-        <p className="sm:min-h-44 md:min-h-48 xl:min-h-36">{description}</p>
-      </div>
-      <Button
-        icon="ArrowUpRight"
-        variant={settings[variant].buttonVariant}
-        className="text-sm"
-        href={`/${locale}/services/${slug}`}
-      >
-        Learn More
-      </Button>
+    <section
+      className={
+        settings[variant].background + " p-8 rounded-lg space-y-8 h-full"
+      }
+    >
+      {icon && (
+        <Icon
+          name={iconify(icon)}
+          className={settings[variant].text + " size-16"}
+        />
+      )}
+      {(title || description) && (
+        <div className="space-y-3">
+          {title && (
+            <h3 className="font-bold text-xl xl:text-2xl text-white sm:min-h-14 lg:min-h-16">
+              {title}
+            </h3>
+          )}
+          {description && (
+            <p className="sm:min-h-44 md:min-h-48 xl:min-h-36">{description}</p>
+          )}
+        </div>
+      )}
+      {cta && (
+        <Button
+          icon="ArrowUpRight"
+          variant={settings[variant].buttonVariant}
+          className="text-sm"
+          href={`/${locale}/services/${slug}`}
+        >
+          {cta}
+        </Button>
+      )}
     </section>
   );
 };