+++ apps/www/src/modules/services/page/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ServicesHeading from "../widgets/heading";
 import { Separator } from "@components/separator";
-import ServiceList from "../widgets/service-list";
+// import ServiceList from "../widgets/service-list";
 import FooterSetup from "@utilities/footer-setup";
 
 const ServicesPage = () => {
@@ -9,7 +9,7 @@ const ServicesPage = () => {
     <>
       <ServicesHeading container />
       <Separator className="mb-28 mt-14" container />
-      <ServiceList container />
+      {/* <ServiceList container /> */}
       <FooterSetup variant="complex" />
     </>
   );